Use numeric denominations in banknote JSON (v2)

+++ src/main/kotlin/atm/api/adapter/in/http/LoadBanknotesController.kt
@@ -34,7 +34,7 @@ data class LoadBanknotesResponse(val message: String)
  */
 fun LoadBanknotesRequest.toCommand(): LoadBanknotesCommand =
     LoadBanknotesCommand(
-        banknotes = banknotes.mapKeys { (note, _) -> Banknote.fromNominal(note.toInt()) }
+        banknotes = banknotes.mapKeys { (note, _) -> Banknote.fromNominal(note) }
     )
 
 /**
